move core whep output logic into whep_output module
